added license headers, tweaked homepage, deleted target files
